chore: :art: Folder and Files renamed according to project standards
